plugins: add test see if the current origin needs setup
ServerBlocks are parsed for each name in the serverblock keys, this may
lead to block being setup multiple times while the intent was probably
to do things once:

~~~
miek.nl miek.org {
    file plugin/sign/testdata/db.miek.nl.signed miek.nl {
        reload 5s
    }
    file plugin/sign/testdata/db.miek.org.signed miek.org {
        reload 5s
    }
}
~~~

Each file plugin should do a thing for the origin specified after it,
but this the first one is done for miek.nl and miek.org and the second
one too; so 4 in total. I don't believe this is the intent.

This add Origins function that checks for this. This returns an
ErrOrigin for which the plugin needs to check. To not fall all tests
that use an bare bones caddy.NewTestController the check is a noop if
the ServerBlockKeys are not there.

Fixes: #3026

Note all plugins that do this need this fix.

Signed-off-by: Miek Gieben <miek@miek.nl>