tests: test kubernetes readme
This adds a way to test the kubernetes readme, it injects a `kubeconfig`
file that points to a close port on localhost. This allows the startup
to continue sufficiently far that the config can actually be tested.

This resulted in tweaks in the README, to enable this testing and a
closer look into the glog/klog stuff that we doing, see #3054 for that.

Signed-off-by: Miek Gieben <miek@miek.nl>